moved min edit distance to its own folder because i use it in a different place, trying to get the malloc warning to go away when running cly.py test. Its somehting to do with how pytest allocated memory, pycoingecko works in interactive shell or when run from the file, but fails in pytest, so annoyting :(
