+++ tests/test_alpaca_account.py
@@ -7,9 +7,6 @@ class TestAlpacaData(unittest.TestCase):
     def setUp(self) -> None:
         self.a = AlpacaAccount()
 
-    def tearDown(self) -> None:
-        del self.a
-
     def test_no_key_id(self):
         try:
             self.a.key_id
