+++ utils/CoinGeckoData.py
@@ -2,7 +2,8 @@ from typing import Dict, List
 import os
 import pandas as pd
 import datetime as date
-from models.settings import Settings
+# from models.settings import Settings
+from utils.Min_Edit_Distance import levenshtein, min_edit_dist
 from pycoingecko import CoinGeckoAPI
 
 
@@ -16,7 +17,13 @@ class CoinGecko:
 
     def get_price(self, symbols: List[str], currency="usd"):
         resp = self.api.get_price(symbols, vs_currencies=currency)
-        print(resp)
+        assert isinstance(symbols, list)
+        if not resp:
+            # got an empty response
+            ...
+            print("No response. Please check the accuracy of your symbols and currencies.")
+            print(self.__find_coin_id(symbols[0], self.get_all_coin_ids()))
+
         return resp
 
     def get_historical_data(self, symbol: str, days: int, currency='usd'):
@@ -39,6 +46,14 @@ class CoinGecko:
         """
         return self.api.get_coins_list()
 
+    def get_all_coin_ids(self) -> set:
+        """IDs are what must be used to get information from pycoingecko"""
+        coins = self.get_coins_list()
+        return set([coin['id'] for coin in coins])
+
+    def __find_coin_id(self, wrongName, allIds):
+        suggestion = min_edit_dist(wrongName, allIds)
+        return f"No data found for symbol '{wrongName}'\n\n Did you mean '{suggestion}'?"
 
 
 c = CoinGecko()
